+++ Dockerfile
@@ -15,7 +15,7 @@ RUN \
   chmod -R 0775 /etc/mysql && \
   chmod -R 0755 /hooks && \
   chmod -R 0777 /var/log/mysql && \
-  cd /opt/configurability/src/ && \
+  cd /opt/configurability/src/mysql_config_translator && \
   pip --no-cache install --upgrade pip && \
   pip --no-cache install --upgrade .
