+++ files/etc/configurability/my.ini
@@ -3,3 +3,4 @@ enabled = true
 module = configurability.ini_merge_process
 ini_file_path = /etc/mysql/my.cnf
 configuration_file_name = configuration-mysql.json
+config_translator_module = mysql_config_translator
