+++ files/opt/configurability/src/mysql_config_translator/__init__.py
@@ -1,0 +1,5 @@
+def process(key, value):
+    if key == 'sql_mode':
+        if isinstance(value, list):
+		    return ','.join(value)
+    return value
